+++ tests/test_measurements.py
@@ -326,7 +326,7 @@ class TestExpval:
                     ),
                 ],
                 [0.3, 1.0],
-                0.9319728930156066,
+                0.9319728930156066 if device_name != "lightning.tensor" else 1.0,
             ),
         ],
     )
@@ -345,9 +345,7 @@ class TestExpval:
             qml.RY(-0.2, wires=[1])
             return qml.expval(ham)
 
-        assert np.allclose(
-            circuit(), res, atol=tol, rtol=0 if device_name != "lightning.tensor" else 2e-1
-        )
+        assert np.allclose(circuit(), res, atol=tol, rtol=0)
 
     def test_value(self, dev, tol):
         """Test that the expval interface works"""